+++ trading_card_game/spec/trading_card_game_spec.rb
@@ -1,46 +1,49 @@
+require_relative '../trading_card_game'
+
 describe 'Trading Card Game' do
-  it 'plays a card that matches the mana cost' do
-    deck = [1]
-    mana = 1
-
-    damage = play_hand(deck, mana)
-    
-    expect(damage).to eql(1)
+  context "Preparation" do
+    it "initial player health" do
+      a_player = Player.new
+      expect(a_player.health).to eq(30)
+    end
   end
 
-  it 'plays a card that inflicts maximum damage' do
-    deck = [2, 1]
-    mana = 1
+  context "Gameplay" do
+    it 'plays a card that matches the mana cost' do
+      deck = [1]
+      mana = 1
 
-    damage = play_hand(deck, mana)
-    
-    expect(damage).to eql(1)
-  end
+      damage = play_hand(deck, mana)
 
-  it 'plays as many cards as to inflict maximum damage' do
-    deck = [2, 1, 1]
-    mana = 3
+      expect(damage).to eql(1)
+    end
 
-    damage = play_hand(deck, mana)
-    
-    expect(damage).to eql(3)
-  end
+    it 'plays a card that inflicts maximum damage' do
+      deck = [2, 1]
+      mana = 1
 
-  it 'plays as many cards as long as mana is not exceeded' do
-    deck = [3, 2, 1]
-    mana = 3
+      damage = play_hand(deck, mana)
 
-    damage = play_hand(deck, mana)
-    
-    expect(damage).to eql(3)
+      expect(damage).to eql(1)
+    end
+
+    it 'plays as many cards as to inflict maximum damage' do
+      deck = [2, 1, 1]
+      mana = 3
+
+      damage = play_hand(deck, mana)
+
+      expect(damage).to eql(3)
+    end
+
+    it 'plays as many cards as long as mana is not exceeded' do
+      deck = [3, 2, 1]
+      mana = 3
+
+      damage = play_hand(deck, mana)
+
+      expect(damage).to eql(3)
+    end
   end
 end
 
-def play_hand(deck, mana)
-  return 0 if deck.empty?
-  head, *tail = deck
-
-  damage = 0
-  damage = head if head <= mana
-  damage + play_hand(tail, mana - damage)
-end
