+++ trading_card_game/trading_card_game.rb
@@ -1,0 +1,15 @@
+class Player
+  def health
+    30
+  end
+end
+
+def play_hand(deck, mana)
+  return 0 if deck.empty?
+  head, *tail = deck
+
+  damage = 0
+  damage = head if head <= mana
+  damage + play_hand(tail, mana - damage)
+end
+
